SSG-12314: Changed response cache age text field in CORS assertion properties dialog to have a max value of Integer.MAX_VALUE instead of Long.MAX_VALUE..
git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@44254 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ modules/gateway/assertions/CORSAssertion/src/main/java/com/l7tech/external/assertions/cors/console/CORSPropertiesDialog.java
@@ -152,7 +152,9 @@ public class CORSPropertiesDialog extends AssertionPropertiesOkCancelSupport<COR
         variablePrefixTextField.setAcceptEmpty(true);
 
         validators = new InputValidator( this, getTitle() );
-        validators.addRule(validators.constrainTextFieldToNumberRange(resourceBundle.getString("responseCacheAgeTitle"), responseCacheAgeTextField,0,Long.MAX_VALUE));
+        validators.addRule(validators.constrainTextFieldToNumberRange(resourceBundle.getString("responseCacheAgeTitle"),
+                responseCacheAgeTextField, 0, Integer.MAX_VALUE));
+
         validators.addRule(new InputValidator.ComponentValidationRule(variablePrefixTextField) {
             @Override
             public String getValidationError() {
